Fix description meta tag for profiles page GH-145
