+++ .changeset/fair-rice-fold.md
@@ -1,0 +1,5 @@
+---
+'@undataforum/gatsby-theme-profiles': patch
+---
+
+Fix description meta tag for profiles page
