+++ themes/gatsby-theme-blog/package.json
@@ -29,12 +29,12 @@
     "react-dom": "^16.9.0"
   },
   "dependencies": {
-    "@undataforum/components": "^0.54.0",
+    "@undataforum/components": "^0.56.0",
     "@undataforum/gatsby-theme-base": "^0.10.0",
     "@undataforum/gatsby-theme-profiles": "^0.4.1",
     "gatsby-image": "^2.2.33",
     "gatsby-plugin-compile-es6-packages": "^2.1.0",
-    "gatsby-source-filesystem": "^2.1.35",
+    "gatsby-source-filesystem": "^2.1.38",
     "prop-types": "^15.7.2",
     "remark": "^11.0.1",
     "strip-markdown": "^3.1.0"
