+++ themes/gatsby-theme-profiles/defaultOptions.js
@@ -1,0 +1,9 @@
+module.exports = {
+  assetPath: 'content/assets/profiles',
+  basePath: '',
+  contentPath: 'content/profiles',
+  description:
+    'Profiles page description placeholder (from @undataforum/gatsby-theme-profiles).',
+  title: 'Profiles',
+  type: 'profile',
+};
