+++ themes/gatsby-theme-profiles/gatsby-node.js
@@ -1,228 +1,4 @@
-const fs = require('fs');
-
-const { createFilePath } = require('gatsby-source-filesystem');
-const remark = require('remark');
-const strip = require('strip-markdown');
-
-const defaultOptions = require('./index');
-
-module.exports.onPreBootstrap = (
-  { reporter },
-  {
-    assetPath = defaultOptions.assetPath,
-    contentPath = defaultOptions.contentPath,
-  }
-) => {
-  const dirs = [assetPath, contentPath];
-  dirs.forEach(dir => {
-    if (fs.existsSync(dir)) return;
-    reporter.info(`Creating directory ${dir}`);
-    fs.mkdirSync(dir, { recursive: true });
-  });
-};
-
-module.exports.sourceNodes = ({ actions, schema }) => {
-  const { createTypes } = actions;
-
-  // Derive Profile nodes from Mdx nodes.
-  // There are 2 types of fields:
-  // 1: fields whose values exists on the Mdx node
-  // 2: fields that only return values after using a resolver to compute a value
-  // See https://www.christopherbiscardi.com/post/constructing-query-types-in-themes
-  createTypes(
-    schema.buildObjectType({
-      name: 'Profile',
-      fields: {
-        id: { type: 'ID!' },
-        // Slug is used as alternative ID to match profiles.
-        slug: { type: 'ID!' },
-        // Type is used to distinguish different profile collections.
-        type: {
-          type: 'String!',
-        },
-        avatar: {
-          type: 'File!',
-          extensions: {
-            fileByRelativePath: {},
-          },
-        },
-        firstName: {
-          type: 'String!',
-        },
-        lastName: {
-          type: 'String!',
-        },
-        name: {
-          type: 'String!',
-        },
-        honorific: {
-          type: 'String',
-        },
-        jobtitle: {
-          type: 'String',
-        },
-        organization: {
-          type: 'String',
-        },
-        description: {
-          type: 'String',
-        },
-        path: {
-          type: 'String!',
-        },
-        // Retrieve resolver from Mdx node and run it.
-        body: {
-          type: 'String!',
-          resolve: async (source, args, context, info) => {
-            const type = info.schema.getType(`Mdx`);
-            const mdxNode = context.nodeModel.getNodeById({
-              id: source.parent,
-            });
-            const resolver = type.getFields().body.resolve;
-            const result = await resolver(mdxNode, args, context, {
-              fieldName: 'body',
-            });
-            return result;
-          },
-        },
-      },
-      interfaces: ['Node'],
-    })
-  );
-};
-
-module.exports.onCreateNode = (
-  { node, actions, getNode, createNodeId, createContentDigest },
-  {
-    basePath = defaultOptions.basePath,
-    contentPath = defaultOptions.contentPath,
-    type = defaultOptions.type,
-  }
-) => {
-  const { createNode, createParentChildLink } = actions;
-
-  // Process MDX nodes only.
-  if (node.internal.type !== `Mdx`) {
-    return;
-  }
-
-  // Parent fileNode makes `name` option from `gatsby-source-filename` available as `sourceInstanceName`.
-  const fileNode = getNode(node.parent);
-  const source = fileNode.sourceInstanceName;
-
-  // Process files in `contentPath` location only.
-  if (source === contentPath) {
-    // Process description.
-    let description;
-    if (node.frontmatter.description) {
-      description = node.frontmatter.description;
-    } else {
-      const match =
-        // Match first paragraph when there are multiple paragraphs
-        // or first paragraph when there is only 1 parapgraph.
-        node.rawBody.match(/\n\n(.+)\n\n/) || node.rawBody.match(/\n\n(.+)\n/);
-      if (match) {
-        // Strip Markdown.
-        description = remark()
-          .use(strip)
-          .processSync(match[1])
-          .contents.trim();
-      }
-    }
-
-    // Process path.
-    let path;
-    if (node.frontmatter.slug) {
-      path = `/${node.frontmatter.slug}/`;
-    } else {
-      // The `basePath` argument is not the same as theme option `basePath`.
-      path = createFilePath({ node, getNode, basePath: contentPath });
-    }
-    // We need slug to create a reproducible ID for foreign key linking.
-    const slug = path.slice(1, -1);
-    // Add theme's basePath.
-    path = `${basePath}${path}`;
-    const profile = {
-      slug,
-      type,
-      // frontmatter.avatar contains relative path.
-      // Type definition links relative path to corresponding File node.
-      avatar: node.frontmatter.avatar,
-      firstName: node.frontmatter.firstName,
-      lastName: node.frontmatter.lastName,
-      name: `${node.frontmatter.firstName} ${node.frontmatter.lastName}`,
-      honorific: node.frontmatter.honorific,
-      jobtitle: node.frontmatter.jobtitle,
-      organization: node.frontmatter.organization,
-      path,
-      description,
-    };
-    const profileNode = {
-      ...profile,
-      // Generated ID is namespaced to plugin.name.
-      id: createNodeId(`${type}-${slug}`),
-      // Make profile node aware of MDX node.
-      parent: node.id,
-      children: [],
-      internal: {
-        type: 'Profile',
-        contentDigest: createContentDigest(profile),
-      },
-    };
-    createNode(profileNode);
-    // Make MDX node aware of derived profile node.
-    createParentChildLink({ parent: node, child: profileNode });
-  }
-};
-
-module.exports.createPages = async (
-  { graphql, actions },
-  {
-    basePath = defaultOptions.basePath,
-    title = defaultOptions.title,
-    description,
-    type = defaultOptions.type,
-  }
-) => {
-  const { createPage } = actions;
-
-  const {
-    data: {
-      allProfile: { nodes: profiles },
-    },
-  } = await graphql(
-    `
-      query($type: String!) {
-        allProfile(filter: { type: { eq: $type } }) {
-          nodes {
-            id
-            path
-          }
-        }
-      }
-    `,
-    { type }
-  );
-
-  // Create individual profile pages.
-  profiles.forEach(({ id, path }) => {
-    createPage({
-      path,
-      component: require.resolve('./src/templates/profile.js'),
-      context: {
-        id,
-      },
-    });
-  });
-
-  // Create profiles page.
-  createPage({
-    path: `${basePath}/`,
-    component: require.resolve('./src/templates/profiles.js'),
-    context: {
-      title,
-      description,
-      type,
-    },
-  });
-};
+module.exports.onPreBootstrap = require('./gatsby/on-pre-bootstrap');
+module.exports.sourceNodes = require('./gatsby/source-nodes');
+module.exports.onCreateNode = require('./gatsby/on-create-node');
+module.exports.createPages = require('./gatsby/create-pages');
