+++ themes/gatsby-theme-profiles/index.js
@@ -1,8 +1,1 @@
-// Default theme options.
-module.exports = {
-  assetPath: 'content/assets/profiles',
-  basePath: '',
-  contentPath: 'content/profiles',
-  title: 'Profiles',
-  type: 'profile',
-};
+// noop
