+++ themes/gatsby-theme-profiles/src/components/profiles-page.js
@@ -40,7 +40,7 @@ ProfilesPage.propTypes = {
     })
   ),
   title: string.isRequired,
-  description: string,
+  description: string.isRequired,
   location: shape({ pathname: string.isRequired }).isRequired,
 };
